refactor goal card
