--- client/src/app/Authorized/PageContent/Goals/GoalCard/EditableGoalNameCell/EditableGoalNameCell.module.css
@@ -1,4 +1,0 @@
-.container {
-  flex: 1 1 auto;
-  align-items: center;
-}
