--- client/src/app/Authorized/PageContent/Goals/GoalCard/EditableGoalTargetAmountCell/EditableGoalTargetAmountCell.module.css
@@ -1,6 +1,0 @@
-.container {
-  flex: 1 1 auto;
-  justify-content: flex-end;
-  align-items: center;
-  gap: 0.25rem;
-}
