--- client/src/app/Authorized/PageContent/Goals/GoalCard/EditableGoalTargetDateCell/EditableGoalTargetDateCell.module.css
@@ -1,5 +1,0 @@
-.container {
-  flex: 1 1 auto;
-  align-items: center;
-  gap: 0.25rem;
-}
